backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
supabase==1.0.3
python-dotenv==1.0.0
gunicorn==20.1.0

